requirements.txt
Django==4.2.5
djangorestframework==3.14.0
psycopg2-binary==2.9.7
isort==5.13.2
python-dotenv==1.0.1